+++ assets/js/components/Application/Application.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import applicationStyles from './Application.scss';
-
-export default class ArrowMenu extends React.Component {
-    render() {
-        return (<div className={applicationStyles.hello}>Hello Zcope</div>);
-    }
-}