+++ assets/js/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import {render} from 'react-dom';
-import Application from './components/Application';
+import Application from './containers/Application';
 
 const applicationElement = document.getElementById('app');