+++ assets/js/views/TimeTracking/TimeTracking.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default class TimeTracking extends React.Component {
+    componentDidMount() {
+        document.title = 'TimeTracking';
+    }
+
+    render() {
+        return (<div>TimeTracking</div>);
+    }
+}